little adjust and move ID operation functions to new file
